+++ package.json
@@ -56,16 +56,16 @@
   },
   "devDependencies": {
     "@fluffy-spoon/substitute": "^1.89.0",
-    "@types/chai": "^4.1.7",
-    "@types/chai-as-promised": "^7.1.1",
-    "@types/debug": "^4.1.4",
-    "@types/express": "^4.16.1",
-    "@types/express-ntlm": "^2.3.0",
+    "@types/chai": "^4.2.0",
+    "@types/chai-as-promised": "^7.1.2",
+    "@types/debug": "^4.1.5",
+    "@types/express": "^4.17.1",
+    "@types/express-ntlm": "^2.3.1",
     "@types/fs-extra": "^8.0.0",
     "@types/mocha": "^5.2.6",
-    "@types/node": "^12.6.9",
-    "@types/node-cleanup": "^2.1.0",
-    "@types/node-forge": "^0.8.5",
+    "@types/node": "^12.7.2",
+    "@types/node-cleanup": "^2.1.1",
+    "@types/node-forge": "^0.8.6",
     "@types/sinon": "^7.0.13",
     "@types/tunnel": "0.0.1",
     "chai": "^4.2.0",
@@ -75,13 +75,13 @@
     "mocha": "^6.2.0",
     "node-forge": "^0.8.5",
     "shx": "^0.3.2",
-    "sinon": "^7.3.2",
+    "sinon": "^7.4.1",
     "source-map-support": "^0.5.13",
     "ts-node": "^8.3.0",
-    "tslint": "^5.18.0",
+    "tslint": "^5.19.0",
     "tslint-eslint-rules": "^5.4.0",
     "tunnel": "0.0.6",
-    "typescript": "^3.5.3",
+    "typescript": "^3.6.2",
     "vrsource-tslint-rules": "^6.0.0"
   },
   "peerDependencies": {
