+++ src/ntlm/ntlm.constants.ts
@@ -1,0 +1,3 @@
+export class NtlmConstants {
+  static readonly NTLM_SIGNATURE = 'NTLMSSP\0';
+}
