+++ src/ntlm/ntlm.message.ts
@@ -1,0 +1,11 @@
+export class NtlmMessage {
+  raw: Buffer = Buffer.alloc(0);
+
+  constructor(buf: Buffer) {
+    this.raw = buf;
+  }
+
+  header(): string {
+    return 'NTLM ' + this.raw.toString('base64');
+  }
+}
