+++ test/ntlm/ntlm.spec.ts
@@ -1,8 +1,16 @@
 import 'mocha';
-const ntlm = require('../../src/ntlm/ntlm');
+
+import { DependencyInjection } from '../../src/proxy/dependency.injection';
+import { TYPES } from '../../src/proxy/dependency.injection.types';
+
+//import Ntlm = require('../../src/ntlm/ntlm');
 import { expect } from 'chai';
+import { INtlm } from '../../src/ntlm/interfaces/i.ntlm';
 
 describe('NTLM hashes', function () {
+  let dependencyInjection = new DependencyInjection();
+  let ntlm: INtlm;
+
   const hostName = 'LightCity';
   const domainName = 'Ursa-Minor';
   const user = 'Zaphod';
@@ -33,24 +41,28 @@ describe('NTLM hashes', function () {
   const NTLMv2asciiType3withFullTargetInfo = 'NTLM TlRMTVNTUAADAAAAGAAYAHEAAADOAM4AiQAAAAoACgBYAAAABgAGAGIAAAAJAAkAaAAAAAAAAABXAQAABgCJAgoAukcAAAAP2fjzqJY/7wxEYeR28AGdg1VSU0EtTUlOT1JaYXBob2RMSUdIVENJVFkAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADXhZHgC0fgPJIzov16QnSMAQEAAAAAAAClLyMZMDDVAUNsaU5vbmNlAAAAAAEAEABNAE8AUwBJAFMATABFAFkAAgAUAFUAUgBTAEEALQBNAEkATgBPAFIAAwAmAE0AbwBzAEkAcwBsAGUAeQAuAHUAcgBzAGEALgBtAGkAbgBvAHIABAAUAHUAcgBzAGEALgBtAGkAbgBvAHIABQAUAHUAcgBzAGEALgBtAGkAbgBvAHIABwAIAKUvIxkwMNUBBgAEAAIAAAAAAAAAAAAAAA==';
   const NTLMv2ucsType3withFullTargetInfo = 'NTLM TlRMTVNTUAADAAAAGAAYAIoAAADOAM4AogAAABQAFABYAAAADAAMAGwAAAASABIAeAAAAAAAAABwAQAABQCJAgoAukcAAAAPZMtzZqis7BLo0lsLULi+9FUAUgBTAEEALQBNAEkATgBPAFIAWgBhAHAAaABvAGQATABJAEcASABUAEMASQBUAFkAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA14WR4AtH4DySM6L9ekJ0jAEBAAAAAAAApS8jGTAw1QFDbGlOb25jZQAAAAABABAATQBPAFMASQBTAEwARQBZAAIAFABVAFIAUwBBAC0ATQBJAE4ATwBSAAMAJgBNAG8AcwBJAHMAbABlAHkALgB1AHIAcwBhAC4AbQBpAG4AbwByAAQAFAB1AHIAcwBhAC4AbQBpAG4AbwByAAUAFAB1AHIAcwBhAC4AbQBpAG4AbwByAAcACAClLyMZMDDVAQYABAACAAAAAAAAAAAAAAA=';
 
+  beforeEach('', function () {
+    ntlm = dependencyInjection.get<INtlm>(TYPES.INtlm);
+  });
+
   it('should create expected Type 1 message. NTLMv1 ASCII', function () {
     const type1 = ntlm.createType1Message(1, '', '');
-    expect(type1).to.be.equal(NTLMv1asciiType1);
+    expect(type1.header()).to.be.equal(NTLMv1asciiType1);
   });
 
   it('should create expected Type 1 message. NTLMv1 ASCII with domain and workstation', function () {
     const type1 = ntlm.createType1Message(1, hostName.toUpperCase(), domainName.toUpperCase());
-    expect(type1).to.be.equal(NTLMv1asciiType1withDomainAndWS);
+    expect(type1.header()).to.be.equal(NTLMv1asciiType1withDomainAndWS);
   });
 
   it('should create expected Type 1 message. NTLMv2 ASCII', function () {
     const type1 = ntlm.createType1Message(2, '', '');
-    expect(type1).to.be.equal(NTLMv2asciiType1);
+    expect(type1.header()).to.be.equal(NTLMv2asciiType1);
   });
 
   it('should create expected Type 1 message. NTLMv2 ASCII with domain and workstation', function () {
     const type1 = ntlm.createType1Message(2, hostName.toUpperCase(), domainName.toUpperCase());
-    expect(type1).to.be.equal(NTLMv2asciiType1withDomainAndWS);
+    expect(type1.header()).to.be.equal(NTLMv2asciiType1withDomainAndWS);
   });
 
   it('should decode Type 2 message. NTLMv1 ASCII', function () {
@@ -120,18 +132,21 @@ describe('NTLM hashes', function () {
   });
 
   it('should create expected Type 3 message. NTLMv1 ASCII', function () {
+    const type1 = ntlm.createType1Message(1, hostName.toUpperCase(), domainName.toUpperCase());
     const type2 = ntlm.decodeType2Message(NTLMv1asciiType2);
-    const type3 = ntlm.createType3Message(type2, user, password, hostName.toUpperCase(), domainName.toUpperCase());
-    expect(type3).to.be.equal(NTLMv1asciiType3);
+    const type3 = ntlm.createType3Message(type1, type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), undefined, undefined);
+    expect(type3.header()).to.be.equal(NTLMv1asciiType3);
   });
 
   it('should create expected Type 3 message. NTLMv1 UCS2', function () {
+    const type1 = ntlm.createType1Message(1, hostName.toUpperCase(), domainName.toUpperCase());
     const type2 = ntlm.decodeType2Message(NTLMv1ucsType2);
-    const type3 = ntlm.createType3Message(type2, user, password, hostName.toUpperCase(), domainName.toUpperCase());
-    expect(type3).to.be.equal(NTLMv1ucsType3);
+    const type3 = ntlm.createType3Message(type1, type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), undefined, undefined);
+    expect(type3.header()).to.be.equal(NTLMv1ucsType3);
   });
 
   it('should create expected Type 3 message. NTLMv2 ASCII', function () {
+    const type1 = ntlm.createType1Message(2, hostName.toUpperCase(), domainName.toUpperCase());
     const type2 = ntlm.decodeType2Message(NTLMv2asciiType2);
     // Since the client nonce and timestamp are generated by the client, we have
     // to force usage of the actual values from our expected answer (which was generated by another the python ntlm-auth client)
@@ -140,11 +155,12 @@ describe('NTLM hashes', function () {
     let timestamp = buf.readUInt32LE(timestamp_index+4).toString(16) +
                     buf.readUInt32LE(timestamp_index).toString(16);
     let client_nonce = buf.slice(timestamp_index+8,timestamp_index+16).toString('hex');
-    const type3 = ntlm.createType3Message(type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), client_nonce, timestamp);
-    expect(type3).to.be.equal(NTLMv2asciiType3);
+    const type3 = ntlm.createType3Message(type1, type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), client_nonce, timestamp);
+    expect(type3.header()).to.be.equal(NTLMv2asciiType3);
   });
 
   it('should create expected Type 3 message. NTLMv2 UCS2', function () {
+    const type1 = ntlm.createType1Message(2, hostName.toUpperCase(), domainName.toUpperCase());
     const type2 = ntlm.decodeType2Message(NTLMv2ucsType2);
     // Since the client nonce and timestamp are generated by the client, we have
     // to force usage of the actual values from our expected answer (which was generated by another the python ntlm-auth client)
@@ -153,30 +169,32 @@ describe('NTLM hashes', function () {
     let timestamp = buf.readUInt32LE(timestamp_index+4).toString(16) +
                     buf.readUInt32LE(timestamp_index).toString(16);
     let client_nonce = buf.slice(timestamp_index+8,timestamp_index+16).toString('hex');
-    const type3 = ntlm.createType3Message(type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), client_nonce, timestamp);
-    expect(type3).to.be.equal(NTLMv2ucsType3);
+    const type3 = ntlm.createType3Message(type1, type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), client_nonce, timestamp);
+    expect(type3.header()).to.be.equal(NTLMv2ucsType3);
   });
 
   it('should create expected Type 3 message. NTLMv2 ASCII with full target info', function () {
+    const type1 = ntlm.createType1Message(2, hostName.toUpperCase(), domainName.toUpperCase());
     const type2 = ntlm.decodeType2Message(NTLMv2asciiType2withFullTargetInfo);
     // Since the client nonce is generated by the client, we have
     // to force usage of the actual values from our expected answer (which was generated by another the python ntlm-auth client)
     let buf = Buffer.from(NTLMv2asciiType3withFullTargetInfo.substring(5), 'base64');
     let timestamp_index = 137+16+8;
     let client_nonce = buf.slice(timestamp_index+8,timestamp_index+16).toString('hex');
-    const type3 = ntlm.createType3Message(type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), client_nonce);
-    expect(type3).to.be.equal(NTLMv2asciiType3withFullTargetInfo);
+    const type3 = ntlm.createType3Message(type1, type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), client_nonce, undefined);
+    expect(type3.header()).to.be.equal(NTLMv2asciiType3withFullTargetInfo);
   });
 
   it('should create expected Type 3 message. NTLMv2 UCS2 with full target info', function () {
+    const type1 = ntlm.createType1Message(2, hostName.toUpperCase(), domainName.toUpperCase());
     const type2 = ntlm.decodeType2Message(NTLMv2ucsType2withFullTargetInfo);
     // Since the client nonce is generated by the client, we have
     // to force usage of the actual values from our expected answer (which was generated by another the python ntlm-auth client)
     let buf = Buffer.from(NTLMv2ucsType3withFullTargetInfo.substring(5), 'base64');
     let timestamp_index = 162+16+8;
     let client_nonce = buf.slice(timestamp_index+8,timestamp_index+16).toString('hex');
-    const type3 = ntlm.createType3Message(type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), client_nonce);
-    expect(type3).to.be.equal(NTLMv2ucsType3withFullTargetInfo);
+    const type3 = ntlm.createType3Message(type1, type2, user, password, hostName.toUpperCase(), domainName.toUpperCase(), client_nonce, undefined);
+    expect(type3.header()).to.be.equal(NTLMv2ucsType3withFullTargetInfo);
   });
 
 });
